+++ tests/conftest.py
@@ -3,6 +3,5 @@ from tests.ui_tests.ui_test_utils import (  # noqa: F401
     template,
     nexus_wrapper,
     instrument,
-    add_component_dialog,
     mock_pixel_options,
 )
