+++ tests/ui_tests/test_ui_fields.py
@@ -1,7 +1,11 @@
 from PySide2.QtWidgets import QListWidget
 import pytest
+
 from nexus_constructor.component_fields import FieldWidget
 from nexus_constructor.stream_fields_widget import StreamFieldsWidget
+from tests.ui_tests.test_ui_add_component_window import (  # noqa: F401
+    add_component_dialog,
+)
 
 
 @pytest.fixture
